+++ admin/admin.go
@@ -33,7 +33,6 @@ import (
 	"encoding/json"
 	"errors"
 	"fmt"
-	"github.com/rudderlabs/rudder-server/router/drain"
 	"net"
 	"net/http"
 	"net/rpc"
@@ -182,34 +181,6 @@ func (a Admin) GetLoggingConfig(noArgs struct{}, reply *string) (err error) {
 	}()
 	loggingConfigMap := logger.GetLoggingConfig()
 	formattedOutput, err := json.MarshalIndent(loggingConfigMap, "", "  ")
-	*reply = string(formattedOutput)
-	return err
-}
-
-func (a Admin) SetDrainJobsConfig(dHandle drain.DrainConfig, reply *string) (err error) {
-	defer func() {
-		if r := recover(); r != nil {
-			pkgLogger.Error(r)
-			err = errors.New("Internal Rudder Server Error")
-		}
-	}()
-
-	_, err = drain.SetDrainJobIDs(dHandle.MinDrainJobID, dHandle.MaxDrainJobID, dHandle.DrainDestinationID)
-	if err == nil {
-		*reply = fmt.Sprintf("Drain config updated")
-	}
-	return err
-}
-
-func (a Admin) GetDrainJobsConfig(noArgs struct{}, reply *string) (err error) {
-	defer func() {
-		if r := recover(); r != nil {
-			pkgLogger.Error(r)
-			err = errors.New("Internal Rudder Server Error")
-		}
-	}()
-	drainHandler := drain.GetDrainJobHandler()
-	formattedOutput, err := json.MarshalIndent(drainHandler, "", "  ")
 	*reply = string(formattedOutput)
 	return err
 }
